requirements.txt

streamlit
langchain
python-dotenv
huggingface_hub
